Rename the notifications package
