+++ WooCommerce/src/main/kotlin/com/woocommerce/android/notifications/LoginHelpNotificationType.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.localnotifications
+package com.woocommerce.android.notifications
 
 enum class LoginHelpNotificationType(private val typeName: String) {
     LOGIN_SITE_ADDRESS_ERROR("site_address_error"),
