+++ WooCommerce/src/main/kotlin/com/woocommerce/android/notifications/LoginHelpNotificationWorker.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.localnotifications
+package com.woocommerce.android.notifications
 
 import android.content.Context
 import android.content.Intent
@@ -8,11 +8,10 @@ import androidx.work.WorkerParameters
 import com.woocommerce.android.AppPrefsWrapper
 import com.woocommerce.android.analytics.AnalyticsEvent.LOGIN_LOCAL_NOTIFICATION_DISPLAYED
 import com.woocommerce.android.analytics.AnalyticsTracker
-import com.woocommerce.android.localnotifications.LoginHelpNotificationType.Companion
 import com.woocommerce.android.push.WooNotificationBuilder
 import com.woocommerce.android.support.help.HelpActivity
 import com.woocommerce.android.support.help.HelpOrigin
-import com.woocommerce.android.localnotifications.LoginNotificationScheduler.Companion.LOGIN_NOTIFICATION_TYPE_KEY
+import com.woocommerce.android.notifications.LoginNotificationScheduler.Companion.LOGIN_NOTIFICATION_TYPE_KEY
 import com.woocommerce.android.viewmodel.ResourceProvider
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
