+++ WooCommerce/src/main/kotlin/com/woocommerce/android/push/WooNotificationBuilder.kt
@@ -17,7 +17,7 @@ import androidx.core.content.ContextCompat
 import com.bumptech.glide.Glide
 import com.woocommerce.android.R
 import com.woocommerce.android.model.Notification
-import com.woocommerce.android.localnotifications.LoginNotificationScheduler
+import com.woocommerce.android.notifications.LoginNotificationScheduler
 import com.woocommerce.android.ui.main.MainActivity
 import com.woocommerce.android.util.SystemVersionUtils
 import com.woocommerce.android.util.WooLog
